echo_requirements.txt
# Echo — Autonomous Brand Strategist
# pip install -r echo_requirements.txt

# Core AI
groq
crewai>=0.28.0
pydantic>=2.0
pydantic-ai

# Web framework
fastapi
uvicorn[standard]
python-dotenv

# Week 1 — Research
requests
firecrawl-py

# Week 1 Day 2 — Vector DB
qdrant-client

# Week 2 — Orchestration
langgraph
langchain
langchain-groq

# Week 2 Day 3 — State persistence
redis

# Week 3 — Observability
langsmith

# Week 4 — Social APIs
tweepy

# Week 4 Day 4 — Scheduler
schedule

# Utilities
httpx
